splits destination exceptions, fixes new schemas in tests
