+++ dlt/load/load.py
@@ -21,15 +21,10 @@ from dlt.common.storages.load_package import LoadPackageStateInjectableContext
 from dlt.common.runners import TRunMetrics, Runnable, workermethod, NullExecutor
 from dlt.common.runtime.collector import Collector, NULL_COLLECTOR
 from dlt.common.runtime.logger import pretty_format_exception
-from dlt.common.exceptions import (
-    TerminalValueError,
-    DestinationTerminalException,
-    DestinationTransientException,
-)
+from dlt.common.exceptions import TerminalValueError
 from dlt.common.configuration.container import Container
-
+from dlt.common.configuration.specs.config_section_context import ConfigSectionContext
 from dlt.common.schema import Schema, TSchemaTables
-
 from dlt.common.storages import LoadStorage
 from dlt.common.destination.reference import (
     DestinationClientDwhConfiguration,
@@ -44,7 +39,10 @@ from dlt.common.destination.reference import (
     SupportsStagingDestination,
     TDestination,
 )
-from dlt.common.configuration.specs.config_section_context import ConfigSectionContext
+from dlt.common.destination.exceptions import (
+    DestinationTerminalException,
+    DestinationTransientException,
+)
 
 from dlt.destinations.job_impl import EmptyLoadJob
